requirements.txt
numpy
requests
rasterio
geopandas
shapely
pyproj
pandas
python-dotenv
